+++ product/agent/agent-lib/src/test/java/se/crisp/codekvast/agent/lib/util/ConfigUtilsTest.java
@@ -56,12 +56,12 @@ public class ConfigUtilsTest {
 
     @Test
     public void testGetNormalizedPrefixes1() throws Exception {
-        assertThat(ConfigUtils.getNormalizedPackagePrefixes("   com.acme... ; foo.bar..   "), hasItems("com.acme", "foo.bar"));
+        assertThat(ConfigUtils.getNormalizedPackages("   com.acme... ; foo.bar..   "), hasItems("com.acme", "foo.bar"));
     }
 
     @Test
     public void testGetNormalizedPrefixes2() throws Exception {
-        assertThat(ConfigUtils.getNormalizedPackagePrefixes(",   , x, : y  ; : com.acme... , foo.bar..  , "),
+        assertThat(ConfigUtils.getNormalizedPackages(",   , x, : y  ; : com.acme... , foo.bar..  , "),
                    hasItems("x", "y", "com.acme", "foo.bar"));
     }