+++ product/agent/daemon/src/main/java/se/crisp/codekvast/agent/daemon/worker/AbstractCollectorDataProcessorImpl.java
@@ -212,10 +212,10 @@ public abstract class AbstractCollectorDataProcessorImpl implements CollectorDat
                       .daemonVersion(daemonConfig.getDaemonVersion())
                       .dumpedAtMillis(jvm.getDumpedAtMillis())
                       .environment(daemonConfig.getEnvironment())
-                      .excludePackagePrefixes(jvm.getCollectorConfig().getNormalizedExcludePackagePrefixes().stream().collect(Collectors.joining(", ")))
+                      .excludePackages(jvm.getCollectorConfig().getNormalizedExcludePackages().stream().collect(Collectors.joining(", ")))
                       .jvmUuid(jvm.getJvmUuid())
                       .methodVisibility(jvm.getCollectorConfig().getMethodFilter().toString())
-                      .packagePrefixes(jvm.getCollectorConfig().getNormalizedPackagePrefixes().stream().collect(Collectors.joining(", ")))
+                      .packages(jvm.getCollectorConfig().getNormalizedPackages().stream().collect(Collectors.joining(", ")))
                       .startedAtMillis(jvm.getStartedAtMillis())
                       .tags(jvm.getCollectorConfig().getTags())
                       .build();